Data-Engineering/3A-Extraer-BM/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
wbgapi
pyarrow